Updated meta_names in excel, added optimization control to fit mod
Added optimization control to fit_mod and phaser
F now is zero for non-fisheries. Was previously unused for surveys but set at the initial parameter values that were not estimated. Now set to zero.
Removed the environmental q bit in the cpp file because it was giving the cod model some trouble.